+++ frontend/src/features/settings/pages/SettingsPage.test.tsx
@@ -4,16 +4,22 @@ import { describe, test } from "vitest";
 import SettingsPage from "./SettingsPage";
 
 describe("SettingsPage", () => {
-  test("should render the settings page", async ({ expect }) => {
+  test("should render the settings page with all sections", async ({
+    expect,
+  }) => {
     const { container } = render(<SettingsPage />);
-
+    const settingsPage = getByTestId<HTMLDivElement>(
+      container,
+      "settings-page",
+    );
     await waitFor(() => {
-      const settingsPage = getByTestId<HTMLDivElement>(
-        container,
-        "settings-page",
-      );
       expect(settingsPage).toBeVisible();
       expect(settingsPage).toHaveTextContent("Django React Starter");
     });
+
+    expect(settingsPage).toHaveTextContent("Preferences");
+    expect(settingsPage).toHaveTextContent("Information");
+    expect(settingsPage).toHaveTextContent("Security");
+    expect(settingsPage).toHaveTextContent("Danger Zone");
   });
 });