+++ frontend/src/tests/mocks/globals.ts
@@ -29,6 +29,17 @@ export const registerGlobalMocks = () => {
       dispatchEvent: vi.fn(),
     }));
 
+  // HTMLDialogElement
+  HTMLDialogElement.prototype.showModal = vi.fn(function (
+    this: HTMLDialogElement,
+  ) {
+    this.setAttribute("open", "");
+  });
+
+  HTMLDialogElement.prototype.close = vi.fn(function (this: HTMLDialogElement) {
+    this.removeAttribute("open");
+  });
+
   // Toastify
   vi.mock("react-toastify", async (importOriginal) => {
     const mod = await importOriginal<typeof import("react-toastify")>();